requirements.txt
soundata
transformers==4.47.1
datasets
evaluate
python-dotenv